feat: get all messages, add other user
This commit introduces the ability to retrieve all messages within a chat room. It also includes the addition of "other user" data, indicating the last read message ID. The changes include new model classes for the result and other user, modified the repository, added other user info to the conversation state and updated UI to show the seen status.