feat: add participant dropdown & refactor create offer screen
This commit introduces a new `ParticipantDropdown` component for managing the number of participants in an offer. It refactors the CreateOfferScreen to use this new component instead of the previous dropdown implementation. Additionally, it ensures that the `Selesai` button is disabled while loading and only enables when form is completed. Also, it updates the `availableUntil` with `selectedDateTimeMillis` to fix create an offer.